Incorporate more code review suggestions
